data_slicing/requirements.txt
boto3==1.28.61
matplotlib==3.7.0
numpy==1.23.5
osgeo==0.0.1
pandas==1.5.3
Pillow==9.4.0
Pillow==10.0.1
rasterio==1.3.8
